+++ ricaun.Revit.DB.Shape.Tests/Utils/ColorExtension.cs
@@ -8,5 +8,15 @@ namespace ricaun.Revit.DB.Shape.Tests.Utils
         {
             return colorA.Red == colorB.Red && colorA.Green == colorB.Green && colorA.Blue == colorB.Blue;
         }
+
+        public static bool ColorEquals(this Color colorA, ColorWithTransparency colorB)
+        {
+            return colorA.Red == colorB.GetRed() && colorA.Green == colorB.GetGreen() && colorA.Blue == colorB.GetBlue();
+        }
+
+        public static bool ColorEquals(this ColorWithTransparency colorA, Color colorB)
+        {
+            return colorB.ColorEquals(colorA);
+        }
     }
 }