+++ ricaun.Revit.DB.Shape/MaterialUtils.cs
@@ -53,22 +53,9 @@ namespace ricaun.Revit.DB.Shape
         /// <param name="red"></param>
         /// <param name="green"></param>
         /// <param name="blue"></param>
-        /// <returns></returns>
-        public static Material CreateMaterial(Document document, byte red, byte green, byte blue)
-        {
-            return CreateMaterial(document, red, green, blue, byte.MaxValue);
-        }
-
-        /// <summary>
-        /// CreateMaterial
-        /// </summary>
-        /// <param name="document"></param>
-        /// <param name="red"></param>
-        /// <param name="green"></param>
-        /// <param name="blue"></param>
         /// <param name="alpha"></param>
         /// <returns></returns>
-        public static Material CreateMaterial(Document document, byte red, byte green, byte blue, byte alpha)
+        public static Material CreateMaterial(Document document, byte red, byte green, byte blue, byte alpha = byte.MaxValue)
         {
             var materialName = MaterialColorName(red, green, blue, alpha);
             var material = FindMaterial(document, materialName);
@@ -87,25 +74,10 @@ namespace ricaun.Revit.DB.Shape
         /// </summary>
         /// <param name="document"></param>
         /// <param name="color"></param>
-        /// <returns></returns>
-        public static Material CreateMaterial(Document document, Color color = null)
-        {
-            if (color is null)
-                color = new Color(128, 128, 128);
-            return CreateMaterial(document, color.Red, color.Green, color.Blue);
-        }
-
-        /// <summary>
-        /// CreateMaterial
-        /// </summary>
-        /// <param name="document"></param>
-        /// <param name="color"></param>
         /// <param name="alpha"></param>
         /// <returns></returns>
-        public static Material CreateMaterial(Document document, Color color, byte alpha)
+        public static Material CreateMaterial(Document document, Color color, byte alpha = byte.MaxValue)
         {
-            if (color is null)
-                color = new Color(128, 128, 128);
             return CreateMaterial(document, color.Red, color.Green, color.Blue, alpha);
         }
 
@@ -165,7 +137,7 @@ namespace ricaun.Revit.DB.Shape
         /// </summary>
         /// <param name="document"></param>
         /// <returns></returns>
-        public static IEnumerable<Material> FindMaterial(Document document)
+        internal static IEnumerable<Material> FindMaterial(Document document)
         {
             // Find materials
             var materials = new FilteredElementCollector(document)
@@ -198,6 +170,12 @@ namespace ricaun.Revit.DB.Shape
             return FindMaterial(document, MaterialColorName(color));
         }
 
+        /// <summary>
+        /// FindMaterial
+        /// </summary>
+        /// <param name="document"></param>
+        /// <param name="color"></param>
+        /// <returns></returns>
         public static Material FindMaterial(Document document, ColorWithTransparency color)
         {
             return FindMaterial(document, MaterialColorName(color));