+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [0.2.0] / 2023-10-01
 ### Shape
 - Create Project `ricaun.Revit.DB.Shape`
+### Shape.Tests
+- Create Project `ricaun.Revit.DB.Shape.Tests`
+- Tests for `MaterialUtils`
+- Tests for `Colors`
 
 ## [0.1.0] / 2023-10-01
 ### Quaternion